requirements.txt
flask
line-bot-sdk
requests
beautifulsoup4
gunicorn

